+++ package/lib/lo_form.dart
@@ -5,6 +5,7 @@ export 'fields/lo_checkbox.dart';
 export 'fields/lo_text_field.dart';
 
 // Core
+export 'src/lo_config.dart';
 export 'src/lo_field.dart';
 export 'src/lo_field_state.dart';
 export 'src/lo_form.dart';